remove the capacity to only add or remove a claim with /api/entities?action=update-claim
letting that responsability to add-claim and remove-claim endpoints
