+++ server/controllers/entities/entities.js
@@ -1,4 +1,5 @@
 import addClaim from '#controllers/entities/add_claim'
+import removeClaim from '#controllers/entities/remove_claim'
 import ActionsControllers from '#lib/actions_controllers'
 import byUrisGet from './by_uris_get.js'
 import contributions from './contributions.js'
@@ -56,6 +57,7 @@ export default {
     authentified: {
       'add-claim': addClaim,
       'update-claim': updateClaim,
+      'remove-claim': removeClaim,
       'update-label': updateLabel,
       'revert-edit': revertEdit,
       'restore-version': restoreVersion,
