+++ server/controllers/entities/remove_claim.js
@@ -1,17 +1,14 @@
+import { claimUpdatersByPrefix } from '#controllers/entities/update_claim'
 import { error_ } from '#lib/error/error'
-// TODO: accept ISBN URIs
-import inv from './lib/update_inv_claim.js'
-import wd from './lib/update_wd_claim.js'
 
 const sanitization = {
   uri: {},
   property: {},
-  'old-value': {},
-  'new-value': {},
+  value: {},
 }
 
 const controller = async (params, req) => {
-  const { uri, property, oldValue, newValue } = params
+  const { uri, property, value } = params
   const [ prefix, id ] = uri.split(':')
 
   const updater = claimUpdatersByPrefix[prefix]
@@ -19,13 +16,8 @@ const controller = async (params, req) => {
     throw error_.new(`unsupported uri prefix: ${prefix}`, 400, uri)
   }
 
-  await updater(req.user, id, property, oldValue, newValue)
+  await updater(req.user, id, property, value, null)
   return { ok: true }
-}
-
-export const claimUpdatersByPrefix = {
-  inv,
-  wd,
 }
 
 export default { sanitization, controller }
